split out tests more
