+++ custom/opm/opm_reports/tests/__init__.py
@@ -2,4 +2,6 @@ from .case_reports import *
 from .test_status import *
 from .test_vhnd import *
 from .test_windows_and_months import *
+from .test_child_conditions import *
+from .test_preg_conditions import *
 # from .old_regression_tests import TestIncentive
