+++ custom/opm/opm_reports/tests/test_preg_conditions.py
@@ -1,0 +1,32 @@
+from datetime import date
+from unittest import TestCase
+
+from .case_reports import Report, Form, OPMCase, MockCaseRow
+
+
+class TestMotherWeightMonitored(TestCase):
+    def setUp(self):
+        self.case = OPMCase(
+            forms=[],
+            edd=date(2014, 10, 15),
+            weight_tri_1="received",
+            weight_tri_2="not_taken",
+       )
+
+    def test_inapplicable_month(self):
+        report = Report(month=7, year=2014, block="Atri")
+        row = MockCaseRow(self.case, report)
+        self.assertEqual(row.preg_month, 7)
+        self.assertEqual(None, row.preg_weighed)
+
+    def test_condition_met(self):
+        report = Report(month=6, year=2014, block="Atri")
+        row = MockCaseRow(self.case, report)
+        self.assertEqual(row.preg_month, 6)
+        self.assertTrue(row.preg_weighed)
+
+    def test_condition_not_met(self):
+        report = Report(month=9, year=2014, block="Atri")
+        row = MockCaseRow(self.case, report)
+        self.assertEqual(row.preg_month, 9)
+        self.assertFalse(row.preg_weighed)
